requirements.txt
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.9.0
websockets>=12.0
pynacl>=1.5.0
ta>=0.11.0
python-dotenv>=1.0.0
structlog>=24.0.0
streamlit>=1.30.0
plotly>=5.18.0
